add grid_size & move voxelization code
